--- eeg_stimulus_project/utils/device_manager.py
@@ -1,4 +1,0 @@
-class DeviceManager: 
-    labrecorder = None
-    lab_recorder_connected = False
-
